add supende page and change payment page
